added: game-pregamemenu TU
Separated code in game TU into its own TU to make code much more
readable.
