+++ src/headers/game-pregamemenu.hpp
@@ -1,0 +1,11 @@
+#ifndef GAMEPREGAMEMENU_H
+#define GAMEPREGAMEMENU_H
+
+namespace Game {
+namespace PreGameSetup {
+void SetUpNewGame();
+void ContinueOldGame();
+} // namespace PreGameSetup
+} // namespace Game
+
+#endif
